requirements.txt
fastapi
uvicorn
firebase-admin
openai
python-dotenv
pydantic
bcrypt
python-jose
